# HTTPServer/http-api-spec.yaml
openapi: 3.0.0
info:
  title: "Game Logic Gateway API"
  description: "클라이언트와 HTTPS 서버를 위한 REST API 명세서"
  version: "1.0.0"
servers:
  - url: http://localhost:3000
    description: "로컬 개발 환경"
paths:
...
  /api/game/match/start:
    post:
      summary: "매치메이킹 시작 요청"
      tags:
        - Game
      security:
        - SessionAuth: [] 
      description: |
        인벤토리 전체 스냅샷을 DB와 대조·갱신한 뒤, warehouse(slot 0~79)를 제외한
        inventory/loadout 영역(slot 80~107)을 loadout으로 확정하여 매치메이킹 대기열에 진입합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameReadyRequest'
      responses:
        '200':
          description: "대기열 진입 성공 (매칭 확인용 티켓 ID 반환)"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                  - properties:
                      data:
                        $ref: '#/components/schemas/MatchTicketData'
        '400':
          description: "잘못된 요청 (유효하지 않은 mapId: ERR_INVALID_MAP_ID, 유효하지 않은 characterType: ERR_INVALID_CHARACTER_TYPE, 형식 오류, 슬롯 중복, CUSTOM 로드아웃 시 무기 미장착: ERR_NO_WEAPON_EQUIPPED)"
        '409':
          description: "인벤토리 스냅샷 불일치 (ERR_SNAPSHOT_MISMATCH) 또는 이미 매칭 중 (ERR_ALREADY_IN_MATCH)"

  # 매치메이킹 상태 확인 (Polling)
  /api/game/match/status:
    get:
      summary: "매치메이킹 상태 확인 (폴링)"
      tags:
        - Game
      security:
        - SessionAuth: []
      description: "발급받은 티켓 ID를 이용해 주기적으로 매칭이 성사되었는지 확인합니다."
      parameters:
        - name: ticketId
          in: query
          required: true
          schema:
            type: string
          description: "매치메이킹 시작 시 발급받은 티켓 ID"
      responses:
        '200':
          description: "상태 조회 성공. 매칭 완료 시 접속 정보가 함께 반환됩니다."
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                  - properties:
                      data:
                        $ref: '#/components/schemas/MatchStatusData'
        '403':
          description: "본인의 티켓이 아님"
        '404':
          description: "존재하지 않거나 만료된 티켓"

  # 매치메이킹 취소
  /api/game/match/cancel:
    post:
      summary: "매치메이킹 취소 요청"
      tags:
        - Game
      security:
        - SessionAuth: []
      description: "대기 중인 매치메이킹을 취소합니다."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ticketId
              properties:
                ticketId:
                  type: string
                  description: "취소할 매칭 티켓 ID"
                  example: "ticket_1234abcd"
      responses:
        '200':
          description: "취소 성공"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
        '403':
          description: "본인의 티켓이 아님"
        '409':
          description: "이미 매칭이 성사되어 취소 불가능"

  # 매치메이킹 서버 접속 정보 획득 (토큰 교환)
  /api/game/match/connect:
    post:
      summary: "인게임 서버 접속 정보 요청 (토큰 교환)"
      tags:
        - Game
      security:
        - SessionAuth: []
      description: "매칭 완료 후 받은 roomToken을 제출하여, 실제 데디케이티드 서버의 IP, Port 및 인게임 인증 키를 발급받습니다."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: 
                - roomToken
              properties:
                roomToken:
                  type: string
                  description: "매칭 성공 시 발급받은 방 입장 토큰"
                  example: "token_abc123"
      responses:
        '200':
          description: "접속 정보 발급 성공"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                  - properties:
                      data:
                        $ref: '#/components/schemas/ConnectResponseData'
        '400':
          description: "잘못된 요청 (토큰 누락)"
        '404':
          description: "유효하지 않거나 이미 사용/만료된 토큰"
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: header
      name: X-Session-Id
      description: "로그인 시 발급받은 세션 키를 입력해주세요."
  schemas:
...
    CommonResponse:
      type: object
      required: [success, code]
      properties:
        success:
          type: boolean
          example: true
        code:
          type: integer
          example: 200
        data:
          type: object
          nullable: true
          description: "API 호출 결과 데이터 (API마다 구조가 다름)"
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      nullable: true
      properties:
        message: { type: string, example: "Invalid request" }
        code: { type: string, example: "ERR_BAD_REQUEST" }

    # ----------------------------------------------------
    # 공통 응답에 담을 진짜 내용물
    # ----------------------------------------------------
...
    InventoryItem:
      type: object
      properties:
        item_id:
          type: integer
          description: "아이템 고유 ID"
          example: 101
        slot_index:
          type: integer
          description: "인벤토리 슬롯 번호"
          example: 0
        quantity:
          type: integer
          description: "보유 수량"
          example: 1
...
    GameReadyRequest:
      type: object
      required:
        - mapId
        - loadoutType
        - characterType
      properties:
        mapId:
          type: integer
          description: "진입할 맵의 식별자 (0: MAP_TUTORIAL, 1: MAP_WINCHESTER)"
          enum: [0, 1]
          example: 0
        characterType:
          type: integer
          description: "선택한 캐릭터 타입 식별자"
          enum: [0, 1, 2]
          example: 0
        loadoutType:
          type: string
          description: "장비 반입 모드 (FREE: 기본 장비 지급, CUSTOM: 내 장비 반입)"
          enum: [FREE, CUSTOM]
          example: "CUSTOM"
        inventory:
          type: array
          description: |
            CUSTOM 모드 필수. 클라이언트 전체 인벤토리 스냅샷.
            slot 0~79: warehouse, slot 80~104: inventory, slot 105~107: loadout.
            warehouse를 제외한 나머지가 인게임 loadout으로 확정됩니다.
          items:
            $ref: '#/components/schemas/InventoryItem'
          example:
            - item_id: 101
              slot_index: 80
              quantity: 5
            - item_id: 105
              slot_index: 105
              quantity: 1
    MatchTicketData:
      type: object
      properties:
        ticketId:
          type: string
          description: "상태 조회에 사용할 매칭 대기표 ID"
          example: "ticket_987654321"
    MatchStatusData:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          description: "현재 매칭 상태 (WAITING: 대기 중 또는 데디서버 생성 중, SUCCESS: 매칭 성공)"
          enum: [WAITING, SUCCESS]
          example: "SUCCESS"
        roomToken:
          type: string
          description: "인게임 진입 시 사용할 방 입장 권한 토큰 (해당 토큰으로 접속 라우팅 처리)"
          example: "room_token_abc"
        mapId:
          type: integer
          description: "매칭된 맵 ID (SUCCESS 상태일 때만 포함)"
          example: 1
...
    ConnectResponseData:
      type: object
      properties:
        ip:
          type: string
          description: "접속해야 할 인게임 UDP 서버 IP"
          example: "127.0.0.1"
        port:
          type: integer
          description: "접속해야 할 인게임 UDP 서버 포트"
          example: 7777
        securityKey:
          type: string
          description: "C++ 데디케이티드 서버가 생성한 숫자형 인증 키 (패킷 통신 시 사용, 2^31 이하)"
          example: "147483647"
        ingameSessionId:
          type: integer
          description: "인게임(UDP) 접속 전용 세션 식별자"
          example: 45